+++ examples/tom.rs
@@ -79,11 +79,11 @@ const PARAM_INFO: [ParamInfo; 8] = [
 struct SharedTom(Arc<Mutex<TomDrum>>);
 
 impl Instrument for SharedTom {
-    fn trigger_with_velocity(&mut self, time: f32, velocity: f32) {
+    fn trigger_with_velocity(&mut self, time: f64, velocity: f32) {
         self.0.lock().unwrap().trigger_with_velocity(time, velocity);
     }
 
-    fn tick(&mut self, current_time: f32) -> f32 {
+    fn tick(&mut self, current_time: f64) -> f32 {
         self.0.lock().unwrap().tick(current_time)
     }
